requirements.txt
# =============================================================================
# ICOM Agent - Python Dependencies
# =============================================================================

# Core
python-dotenv==1.1.0
pydantic==2.11.1

# Database
sqlalchemy==2.0.40

# Data Processing
pandas==2.2.3
numpy==2.2.4
openpyxl==3.1.5

# ML / Prediction (Phase 0)
scikit-learn==1.6.1
xgboost==3.0.0
# prophet==1.1.6  # Install separately: pip install prophet

# Visualization (Phase 0 EDA)
matplotlib==3.10.1
seaborn==0.13.2
jupyter==1.1.1

# API Server (Phase 1)
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1

# Dashboard (Phase 1)
streamlit==1.44.1

# LLM / Agent (Phase 2-3)
# langchain==0.3.x
# langgraph==0.3.x
# openai==1.x

# Scheduling
apscheduler==3.11.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
